fix(routing): canonical post URLs and forum board thread listing
Consolidate post detail routes to /b/{board}/{id}/{slug} — the old
/threads/ and /feed/ prefixed routes now 301 redirect to the canonical
path. Update postUrl() helper and all inline URL generation to match.

Fix forum board index showing blank feed by switching useBoard to
useState so parent layout and child pages share board data. Previously
the child created a fresh useBoard() instance with null board, causing
the postType filter to default to 'feed' and filter out all threads.
