+++ frontend/pages/@[username]/comments.vue
@@ -103,7 +103,7 @@ await fetchComments()
           <span class="text-gray-400">commented on</span>
           <NuxtLink
             v-if="comment.post"
-            :to="`/b/${comment.post.board?.name || comment.board?.name || 'unknown'}/feed/${comment.postId}/${comment.post.slug || ''}`"
+            :to="`/b/${comment.post.board?.name || comment.board?.name || 'unknown'}/${comment.postId}/${comment.post.slug || ''}`"
             class="font-medium text-gray-700 no-underline hover:text-primary truncate"
           >
             {{ comment.post.title }}
